refactor(types): rename DatasetV2 to Dataset, consolidate provenance file
- datasetV2.ts -> dataset.ts, DatasetV2 interface -> Dataset
- DatasetV2Provenance.tsx merged into DatasetProvenance.tsx (wrapper deleted)
- Update all import sites
