+++ src/js/features/beacon/utils.ts
@@ -12,7 +12,7 @@ import type { SearchFieldAndOptions, SearchFieldSection } from '@/features/searc
 import type { ChartData, DiscoveryResults, OptionalDiscoveryResults } from '@/types/data';
 import type { NetworkBeacon } from '@/types/beaconNetwork';
 import type { Project } from '@/types/metadata';
-import type { DatasetV2 } from '@/types/datasetV2';
+import type { Dataset } from '@/types/dataset';
 import { RequestStatus } from '@/types/requests';
 import { serializeChartData } from '@/utils/chart';
 
@@ -110,7 +110,7 @@ const scopedBeaconBaseUrl = (projectId: Project['identifier'] | undefined): stri
 
 export const scopedBeaconFilteringTermsUrl = (
   projectId: Project['identifier'] | undefined,
-  datasetId: DatasetV2['identifier'] | undefined
+  datasetId: Dataset['identifier'] | undefined
 ): string => {
   const datasetIdParam = datasetId ? '?' + 'datasetIds=' + datasetId : '';
   return scopedBeaconBaseUrl(projectId) + BEACON_FILTERING_TERMS_PATH + datasetIdParam;
